# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  requestdto.AddMarket:
    properties:
      competitionId:
        description: IPL / BBL
        type: string
      eventId:
        description: Unique Event Id
        type: string
      marketId:
        description: Unique Market Id
        type: string
      marketType:
        description: MATCH_ODDS / BOOKMAKER / FANCY
        type: string
      operatorId:
        type: string
      providerId:
        description: Dream / BetFair / SportRadar
        type: string
      sportId:
        description: 1 / 2 / 3
        type: string
    type: object
...
  requestdto.GetCacheReq:
    properties:
      cacheType:
        type: string
      key:
        type: string
    type: object
...
  responsedto.DefaultRespDto:
    properties:
      errorDescription:
        type: string
      status:
        type: string
    type: object
...
  responsedto.GetCacheResp:
    properties:
      errorDescription:
        type: string
      respJson:
        type: string
      status:
        type: string
    type: object
...
info:
  contact:
    email: amit.m@outlook.com
    name: Amit
  description: This is an API Documentation for Sports Aggregation Platform (SAP)
  title: SAP
  version: "1.0"
paths:
  /cache/get-cache:
    post:
      consumes:
      - application/json
      description: Endpoint to read cache data
      parameters:
      - description: GetCacheReq model is used
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/requestdto.GetCacheReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsedto.GetCacheResp'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/responsedto.GetCacheResp'
      summary: Get Cache Data
      tags:
      - Cache-Service
...
  /feed/add-market:
    post:
      consumes:
      - application/json
      description: To notify SAP to add Market in the system if not present. This
        helps to maintain the market status
      parameters:
      - description: AddMarket model is used
        in: body
        name: AddMarket
        required: true
        schema:
          $ref: '#/definitions/requestdto.AddMarket'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsedto.DefaultRespDto'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/responsedto.DefaultRespDto'
      summary: Add Market API
      tags:
      - Feed-Service
...
swagger: "2.0"